requirements.txt
# Core dependencies for Collagen webhook processor
cloudinary>=1.36.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
